+++ src/mix/types.ts
@@ -17,7 +17,7 @@ import {BuildsParams, BuildsSearchParams} from './api/builds-types'
 import {DataHostsParams, DataHostsSearchParams} from './api/data-hosts-types'
 import {DeploymentFlowsParams, DeploymentFlowsSearchParams} from './api/deployment-flows-types'
 import {EnginePacksParams} from './api/engine-packs-types'
-import {EntitiesParams} from './api/entities-types'
+import {EntitiesParams, EntitiesSearchParams} from './api/entities-types'
 import {EnvironmentsParams, EnvironmentsSearchParams} from './api/environments-types'
 import {GeographiesParams, GeographiesSearchParams} from './api/geographies-types'
 import {JobsParams, JobsSearchParams} from './api/jobs-types'
@@ -75,6 +75,7 @@ export type MixRequestSearchParams =
   | BuildsSearchParams
   | DataHostsSearchParams
   | DeploymentFlowsSearchParams
+  | EntitiesSearchParams
   | EnvironmentsSearchParams
   | GeographiesSearchParams
   | JobsSearchParams